Move code into src
Move more code into src
