+++ examples/hybrid/driver.jl
@@ -1,32 +1,32 @@
 import ClimaAtmos as CA
 
-include("cli_options.jl")
-(s, parsed_args_defaults) = parse_commandline()
+(s, parsed_args_defaults) = CA.parse_commandline()
 if !(@isdefined parsed_args)
     parsed_args = parsed_args_defaults
 end
 
-include("nvtx.jl")
+# TODO: can we move this into src/?
+using NVTX
+using ClimaComms
+if NVTX.isactive()
+    NVTX.enable_gc_hooks()
+    # makes output on buildkite a bit nicer
+    if ClimaComms.iamroot(comms_ctx)
+        atexit() do
+            println("--- Saving profiler information")
+        end
+    end
+end
 
 parse_arg(pa, key, default) = isnothing(pa[key]) ? default : pa[key]
 
 const FT = parsed_args["FLOAT_TYPE"] == "Float64" ? Float64 : Float32
 
 include("parameter_set.jl")
-params, parsed_args = create_parameter_set(FT, parsed_args, cli_defaults(s))
+params, parsed_args = create_parameter_set(FT, parsed_args, CA.cli_defaults(s))
 
 include("comms.jl")
-if startswith(parsed_args["ode_algo"], "ODE.") # TODO: use Preferences.jl instead:
-    include("../ordinary_diff_eq_bug_fixes.jl")
-end
 fps = parsed_args["fps"]
-idealized_insolation = parsed_args["idealized_insolation"]
-idealized_clouds = parsed_args["idealized_clouds"]
-
-@assert idealized_insolation in (true, false)
-@assert idealized_clouds in (true, false)
-@assert parsed_args["config"] in ("sphere", "column", "box", "plane")
-
 import ClimaAtmos.RRTMGPInterface as RRTMGPI
 import ClimaAtmos.InitialConditions as ICs
 
@@ -37,8 +37,7 @@ import ClimaAtmos.TurbulenceConvection as TC
 atmos = CA.get_atmos(FT, parsed_args, params.turbconv_params)
 @info "AtmosModel: \n$(summary(atmos))"
 numerics = CA.get_numerics(parsed_args)
-include("get_simulation_and_args_integrator.jl")
-simulation = get_simulation(FT, parsed_args)
+simulation = CA.get_simulation(FT, parsed_args)
 initial_condition = CA.get_initial_condition(parsed_args)
 
 # TODO: use import instead of using
@@ -54,106 +53,6 @@ using ClimaTimeSteppers
 
 import Random
 Random.seed!(1234)
-
-# TODO: flip order so that NamedTuple() is fallback.
-function additional_cache(Y, default_cache, parsed_args, params, atmos, dt)
-    (; precip_model, forcing_type, radiation_mode, turbconv_model) = atmos
-
-    radiation_cache = if radiation_mode isa RRTMGPI.AbstractRRTMGPMode
-        CA.radiation_model_cache(
-            Y,
-            default_cache,
-            params,
-            radiation_mode;
-            idealized_insolation,
-            idealized_clouds,
-            data_loader = CA.rrtmgp_data_loader,
-        )
-    else
-        CA.radiation_model_cache(Y, params, radiation_mode)
-    end
-
-    return merge(
-        CA.rayleigh_sponge_cache(atmos.rayleigh_sponge, Y),
-        CA.viscous_sponge_cache(atmos.viscous_sponge, Y),
-        CA.precipitation_cache(Y, precip_model),
-        CA.subsidence_cache(Y, atmos.subsidence),
-        CA.large_scale_advection_cache(Y, atmos.ls_adv),
-        CA.edmf_coriolis_cache(Y, atmos.edmf_coriolis),
-        CA.forcing_cache(Y, forcing_type),
-        radiation_cache,
-        CA.vertical_diffusion_boundary_layer_cache(Y, atmos),
-        CA.non_orographic_gravity_wave_cache(
-            atmos.non_orographic_gravity_wave,
-            atmos.model_config,
-            Y,
-        ),
-        CA.orographic_gravity_wave_cache(atmos.orographic_gravity_wave, Y),
-        (; Δt = dt),
-        CA.turbconv_cache(
-            Y,
-            turbconv_model,
-            atmos,
-            params,
-            parsed_args,
-            initial_condition,
-        ),
-    )
-end
-
-function additional_tendency!(Yₜ, Y, p, t)
-    CA.viscous_sponge_tendency!(Yₜ, Y, p, t, p.atmos.viscous_sponge)
-
-    # Vertical tendencies
-    Fields.bycolumn(axes(Y.c)) do colidx
-        CA.rayleigh_sponge_tendency!(
-            Yₜ,
-            Y,
-            p,
-            t,
-            colidx,
-            p.atmos.rayleigh_sponge,
-        )
-        CA.forcing_tendency!(Yₜ, Y, p, t, colidx, p.forcing_type)
-        CA.subsidence_tendency!(Yₜ, Y, p, t, colidx, p.subsidence)
-        CA.edmf_coriolis_tendency!(Yₜ, Y, p, t, colidx, p.edmf_coriolis)
-        CA.large_scale_advection_tendency!(Yₜ, Y, p, t, colidx, p.ls_adv)
-
-        (; vert_diff) = p.atmos
-        if p.atmos.coupling isa CA.Decoupled
-            CA.get_surface_fluxes!(Y, p, t, colidx, vert_diff)
-        end
-        CA.vertical_diffusion_boundary_layer_tendency!(
-            Yₜ,
-            Y,
-            p,
-            t,
-            colidx,
-            vert_diff,
-        )
-
-        CA.radiation_tendency!(Yₜ, Y, p, t, colidx, p.radiation_model)
-        CA.explicit_sgs_flux_tendency!(Yₜ, Y, p, t, colidx, p.turbconv_model)
-        CA.precipitation_tendency!(Yₜ, Y, p, t, colidx, p.precip_model)
-    end
-    # TODO: make bycolumn-able
-    CA.non_orographic_gravity_wave_tendency!(
-        Yₜ,
-        Y,
-        p,
-        t,
-        p.atmos.non_orographic_gravity_wave,
-    )
-    CA.orographic_gravity_wave_tendency!(
-        Yₜ,
-        Y,
-        p,
-        t,
-        p.atmos.orographic_gravity_wave,
-    )
-end
-
-################################################################################
 
 using OrdinaryDiffEq
 using DiffEqCallbacks
@@ -172,8 +71,6 @@ import ClimaAtmos.TurbulenceConvection as TC
 import ClimaCore.Operators as CCO
 const CM = CloudMicrophysics
 import ClimaAtmos.Parameters as CAP
-
-include("get_cache_and_tendency.jl")
 
 import ClimaCore: enable_threading
 const enable_clima_core_threading = parsed_args["enable_threading"]
@@ -194,7 +91,7 @@ else
 end
 
 @time "Allocating cache (p)" begin
-    p = get_cache(
+    p = CA.get_cache(
         Y,
         parsed_args,
         params,
@@ -202,6 +99,7 @@ end
         atmos,
         numerics,
         simulation,
+        initial_condition,
         comms_ctx,
     )
 end
@@ -212,13 +110,11 @@ end
 
 @time "ode_configuration" ode_algo = CA.ode_configuration(Y, parsed_args, atmos)
 
-include("get_callbacks.jl")
-
 @time "get_callbacks" callback =
-    get_callbacks(parsed_args, simulation, atmos, params)
+    CA.get_callbacks(parsed_args, simulation, atmos, params)
 tspan = (t_start, simulation.t_end)
 @time "args_integrator" integrator_args, integrator_kwargs =
-    args_integrator(parsed_args, Y, p, tspan, ode_algo, callback)
+    CA.args_integrator(parsed_args, Y, p, tspan, ode_algo, callback)
 
 if haskey(ENV, "CI_PERF_SKIP_INIT") # for performance analysis
     throw(:exit_profile_init)
