+++ perf/config_parsed_args.jl
@@ -18,8 +18,9 @@ push!(ARGS, "--rad", "nothing");
 include("perf/config_parsed_args.jl")
 ```
 =#
-ca_dir = joinpath(dirname(@__DIR__));
-include(joinpath(ca_dir, "examples", "hybrid", "cli_options.jl"))
+ca_dir = dirname(@__DIR__)
+include(joinpath(ca_dir, "src", "utils", "cli_options.jl"));
+include(joinpath(ca_dir, "src", "utils", "yaml_helper.jl"))
 (s, _parsed_args) = parse_commandline()
 parsed_args_defaults = cli_defaults(s);
 dict = parsed_args_per_job_id(; filter_name = "driver.jl");
